# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Qareeb - charity platform connecting helpers and people in need
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPayload"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/requests:
    get:
      operationId: adminListRequests
      tags: [admin]
      summary: List all requests (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, fulfilled, rejected]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HelpRequestWithUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/requests/{id}/status:
    patch:
      operationId: adminUpdateRequestStatus
      tags: [admin]
      summary: Approve / reject / fulfill a request (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusPayload"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/requests/{id}:
    delete:
      operationId: adminDeleteRequest
      tags: [admin]
      summary: Delete a request (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWithStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/trust:
    patch:
      operationId: adminUpdateUserTrust
      tags: [admin]
      summary: Update a user's trust status (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrustPayload"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: List reports (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportWithRequest"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        trustStatus:
          type: string
          enum: [new, trusted, restricted]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, trustStatus, createdAt]
    UserWithStats:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        trustStatus:
          type: string
          enum: [new, trusted, restricted]
        createdAt:
          type: string
          format: date-time
        requestsCount:
          type: integer
        helpActionsCount:
          type: integer
      required: [id, name, email, role, trustStatus, createdAt, requestsCount, helpActionsCount]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    RegisterPayload:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
      required: [name, email, password]
    LoginPayload:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    HelpRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        displayName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [food, medicine, clothes, rent, other]
        city:
          type: string
        urgency:
          type: string
          enum: [low, medium, high]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, fulfilled, rejected]
        helpersCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, description, category, city, urgency, status, helpersCount, createdAt]
    HelpRequestWithUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        displayName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [food, medicine, clothes, rent, other]
        city:
          type: string
        urgency:
          type: string
          enum: [low, medium, high]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, fulfilled, rejected]
        helpersCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        userName:
          type: string
        userEmail:
          type: string
        userTrustStatus:
          type: string
          enum: [new, trusted, restricted]
      required: [id, userId, title, description, category, city, urgency, status, helpersCount, createdAt, userName, userEmail, userTrustStatus]
...
    ReportWithRequest:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        reporterUserId:
          type: integer
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        requestTitle:
          type: string
        reporterName:
          type: string
      required: [id, requestId, reporterUserId, reason, createdAt, requestTitle, reporterName]
...
    UpdateStatusPayload:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, fulfilled, rejected]
      required: [status]
    UpdateTrustPayload:
      type: object
      properties:
        trustStatus:
          type: string
          enum: [new, trusted, restricted]
      required: [trustStatus]
...
    AdminStats:
      type: object
      properties:
        pendingRequests:
          type: integer
        approvedRequests:
          type: integer
        fulfilledRequests:
          type: integer
        rejectedRequests:
          type: integer
        totalUsers:
          type: integer
        totalReports:
          type: integer
        totalHelpActions:
          type: integer
      required: [pendingRequests, approvedRequests, fulfilledRequests, rejectedRequests, totalUsers, totalReports, totalHelpActions]
...