# lib/api-zod/src/generated/types/registerPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qareeb - charity platform connecting helpers and people in need
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterPayload {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  /** @minLength 6 */
  password: string;
}