# lib/api-zod/src/generated/types/reportWithRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qareeb - charity platform connecting helpers and people in need
 * OpenAPI spec version: 0.1.0
 */

export interface ReportWithRequest {
  id: number;
  requestId: number;
  reporterUserId: number;
  reason: string;
  createdAt: Date;
  requestTitle: string;
  reporterName: string;
}