# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qareeb - charity platform connecting helpers and people in need
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserTrustStatus } from "./userTrustStatus";

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  trustStatus: UserTrustStatus;
  createdAt: Date;
}