# lib/api-zod/src/generated/types/userWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Qareeb - charity platform connecting helpers and people in need
 * OpenAPI spec version: 0.1.0
 */
import type { UserWithStatsRole } from "./userWithStatsRole";
import type { UserWithStatsTrustStatus } from "./userWithStatsTrustStatus";

export interface UserWithStats {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserWithStatsRole;
  trustStatus: UserWithStatsTrustStatus;
  createdAt: Date;
  requestsCount: number;
  helpActionsCount: number;
}